Add --noconfirm to install_root.sh

+++ root/install_root.sh
@@ -3,6 +3,12 @@
 # Set workdir
 cd "$(dirname "$0")"
 
+# Argument
+NO_CONFIRM=False
+if [[ $1 == "--noconfirm" ]]; then
+    NO_CONFIRM=True
+fi
+
 # List all files
 files=$(find . -type f)
 
@@ -20,6 +26,11 @@ for file in "${splitMultiChar[@]}"; do
     file=$(echo $file|tr -d '\n')
     if [[ "$file" != "" && "$file" != "install_root.sh" ]]; then
         printf "\n./$file --> /$file\n"
-        sudo cp -i "./$file" "/$file"
+        if [[ $NO_CONFIRM == True ]]; then
+            sudo cp "./$file" "/$file"
+        else
+            sudo cp -i "./$file" "/$file"
+        fi
+        
     fi
 done
